add casing subpath export
