+++ packages/compiler/src/utils/index.ts
@@ -3,12 +3,5 @@
 // Be explicit about what get exported so we don't export utils that are not meant to be public.
 // ---------------------------------------
 export { DuplicateTracker } from "./duplicate-tracker.js";
-export {
-  Queue,
-  TwoLevelMap,
-  capitalize,
-  createRekeyableMap,
-  deepClone,
-  deepEquals,
-} from "./misc.js";
+export { Queue, TwoLevelMap, createRekeyableMap, deepClone, deepEquals } from "./misc.js";
 export { useStateMap, useStateSet } from "./state-accessor.js";
