+++ packages/compiler/src/utils/misc.ts
@@ -480,10 +480,3 @@ class RekeyableMapImpl<K, V> implements RekeyableMap<K, V> {
     return true;
   }
 }
-
-/**
- * Simple utility function to capitalize a string.
- */
-export function capitalize<S extends string>(s: S): Capitalize<S> {
-  return (s.charAt(0).toUpperCase() + s.slice(1)) as Capitalize<S>;
-}
